Move DllImportResolver code to ModuleInitializer
Mono does not run the static constructor for NativeMethods before looking up the library for the first P/Invoke method called.
